Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int r;
+    float r;
     scanf("%f",&r);
     float pi=3.14;
     float area=pi*r*r;
